requirements.txt
colorama>=0.4.6
pyfiglet>=0.8.post1
requests>=2.31.0
dnspython>=2.4.2
urllib3>=2.0.7
tqdm>=4.66.1
PyYAML>=6.0.1 